Tests for the BellMaster and some corrections

+++ src/main/java/de/nhilbert/zenbell/de/nhilbert/zenbell/model/Settings.java
@@ -30,5 +30,5 @@ public interface Settings {
 
     void disableNotification();
 
-    long getPeriod();
+    Integer getPeriod();
 }
